+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ### Added
 
-- Added `Installation.objects.acreate_from_gh_data`/`Installation.objects.create_from_gh_data` manager methods.
+- Added `acreate_from_gh_data`/`create_from_gh_data` manager methods to `Installation` and `Repository` models.
 
 ## [0.1.0]