+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.2.0]
+
 ### Added
 
 - Added `acreate_from_gh_data`/`create_from_gh_data` manager methods to `Installation` and `Repository` models.
@@ -46,5 +48,6 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/joshuadavidthomas/django-github-app/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/joshuadavidthomas/django-github-app/compare/v0.2.0...HEAD
 [0.1.0]: https://github.com/joshuadavidthomas/django-github-app/releases/tag/v0.1.0
+[0.2.0]: https://github.com/joshuadavidthomas/django-github-app/releases/tag/v0.2.0